refactor: new location to hooks folder
