+++ src/components/banner.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useTheme } from '@/app/hooks/use-theme'
+import { useTheme } from '@/hooks/use-theme'
 import { CSSProperties } from 'react'
 
 export const Banner = () => {
