+++ src/components/menu.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { useTheme } from '@/app/hooks/use-theme'
+import { useTheme } from '@/hooks/use-theme'
 import { CSSProperties } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
